Improve magnetic fields
feat: 'geo' type magnetic field.
Add function which writes magnetic field grid in the shape of a
rectangualr prism. Use this function rather than creating a
text file separately and then reading it in simCRpropa.
This eliminates the need for the 'txt' magnetic field type and
introduces the 'geo' type, which will be used for various geometries.
The 'txt' type code is still included as it may be usd in the future
for more complicated geometries.
Introduces a new type 'geo' which means filepaths must be updated.
feat: log random seed used for vector direction.
feat: replace statements which autofill when the key is not in the
config with breaks if the key is not in the config, to ensure the user
is aware always of what is used.